Fix elinac's scheme to run it with Taurus 4
In elinac example there was an old scheme 'taco', which is incompatible with Taurus. Taurus 3 couldn't find example's scheme name, because of functional limitation, and was always setting it to the default scheme 'Tango'. Taurus 4 has improved functionality and can find scheme name properly, but during running the vaccagui synoptic panel there was being showed an error that scheme 'taco' is invalid.

:a